train_azure/requirements.txt
azureml.core
